+++ kaskade/kafka/mappers.py
@@ -1,9 +1,11 @@
 import json
 
+from confluent_kafka import Node as NodeMetadata
 from confluent_kafka import TopicPartition as GroupPartitionMetadata
 from confluent_kafka.admin import BrokerMetadata
-from confluent_kafka.admin import GroupMember as GroupMemberMetadata
-from confluent_kafka.admin import GroupMetadata, PartitionMetadata, TopicMetadata
+from confluent_kafka.admin import ConsumerGroupDescription as GroupMetadata
+from confluent_kafka.admin import MemberDescription as MemberMetadata
+from confluent_kafka.admin import PartitionMetadata, TopicMetadata
 from confluent_kafka.schema_registry import Schema as SchemaMetadata
 
 from kaskade.kafka.models import (
@@ -11,6 +13,7 @@ from kaskade.kafka.models import (
     Group,
     GroupMember,
     GroupPartition,
+    Node,
     Partition,
     Schema,
     Topic,
@@ -21,31 +24,41 @@ def metadata_to_broker(metadata: BrokerMetadata) -> Broker:
     return Broker(id=metadata.id, host=metadata.host, port=metadata.port)
 
 
+def metadata_to_node(metadata: NodeMetadata) -> Node:
+    return Node(
+        id=metadata.id, host=metadata.host, port=metadata.port, rack=metadata.rack
+    )
+
+
 def metadata_to_group(metadata: GroupMetadata) -> Group:
     return Group(
-        id=metadata.id,
-        broker=metadata_to_broker(metadata.broker),
-        state=metadata.state,
+        id=metadata.group_id,
+        broker=metadata_to_node(metadata.coordinator),
+        state=str(metadata.state),
+        partition_assignor=metadata.partition_assignor,
         members=[],
         partitions=[],
     )
 
 
-def metadata_to_group_member(metadata: GroupMemberMetadata) -> GroupMember:
+def metadata_to_group_member(group: str, metadata: MemberMetadata) -> GroupMember:
     return GroupMember(
-        id=metadata.id,
-        group="",
+        id=metadata.member_id,
+        group=group,
         client_id=metadata.client_id,
-        client_host=metadata.client_host,
+        host=metadata.host,
+        instance_id=metadata.group_instance_id,
     )
 
 
-def metadata_to_group_partition(metadata: GroupPartitionMetadata) -> GroupPartition:
+def metadata_to_group_partition(
+    group: str, metadata: GroupPartitionMetadata
+) -> GroupPartition:
     return GroupPartition(
         id=metadata.partition,
         topic=metadata.topic,
         offset=metadata.offset,
-        group="",
+        group=group,
         high=0,
         low=0,
     )